// src/book/dto/create-book.dto.ts
import { IsNotEmpty, IsNumber, IsString } from 'class-validator';

interface data {
  sections: number;
  pages: number;
  price: string;
  version: number;
}

export class CreateBookDto {
  @IsString()
  @IsNotEmpty()
  name: string;

  @IsString()
  @IsNotEmpty()
  body: string;

  @IsNotEmpty()
  data: data;

  public?: boolean;

  category?: string;

  @IsNumber()
  @IsNotEmpty()
  id: number;

  @IsString()
  @IsNotEmpty()
  password: string;
}